--- a/src/daft-catalog/src/lib.rs
+++ b/src/daft-catalog/src/lib.rs
@@ -198,7 +198,6 @@ mod tests {
             Arc::new(SourceInfo::PlaceHolder(PlaceHolderInfo {
                 source_schema: schema,
                 clustering_spec: Arc::new(ClusteringSpec::unknown()),
-                source_id: 0,
             })),
         ))
         .arced()