--- a/src/daft-logical-plan/src/ops/source.rs
+++ b/src/daft-logical-plan/src/ops/source.rs
@@ -116,12 +116,9 @@ impl Source {
                 res.push(format!("Number of partitions = {}", num_partitions));
             }
             SourceInfo::PlaceHolder(PlaceHolderInfo {
-                source_id,
-                clustering_spec,
-                ..
+                clustering_spec, ..
             }) => {
                 res.push("PlaceHolder:".to_string());
-                res.push(format!("Source ID = {}", source_id));
                 res.extend(clustering_spec.multiline_display());
             }
         }