--- a/src/daft-logical-plan/src/optimization/rules/simplify_expressions.rs
+++ b/src/daft-logical-plan/src/optimization/rules/simplify_expressions.rs
@@ -70,7 +70,6 @@ mod test {
                 source_info: Arc::new(SourceInfo::PlaceHolder(PlaceHolderInfo {
                     source_schema: schema,
                     clustering_spec: Arc::new(ClusteringSpec::unknown()),
-                    source_id: 0,
                 })),
                 stats_state: StatsState::NotMaterialized,
             })