--- a/src/daft-physical-plan/src/ops/in_memory.rs
+++ b/src/daft-physical-plan/src/ops/in_memory.rs
@@ -34,8 +34,8 @@ impl InMemoryScan {
             "Clustering spec = {{ {} }}",
             self.clustering_spec.multiline_display().join(", ")
         ));
-        if let Some(source_id) = self.in_memory_info.source_id {
-            res.push(format!("Source ID = {}", source_id));
+        if let Some(source_stage_id) = self.in_memory_info.source_stage_id {
+            res.push(format!("Source Stage ID = {}", source_stage_id));
         }
         res
     }
@@ -50,13 +50,13 @@ impl TreeDisplay for InMemoryScan {
 Schema = {},
 Size bytes = {},
 Clustering spec = {{ {} }}
-Source ID = {}
+Source Stage ID = {}
 ",
                     self.schema.short_string(),
                     self.in_memory_info.size_bytes,
                     self.clustering_spec.multiline_display().join(", "),
-                    match self.in_memory_info.source_id {
-                        Some(source_id) => source_id.to_string(),
+                    match self.in_memory_info.source_stage_id {
+                        Some(source_stage_id) => source_stage_id.to_string(),
                         None => "None".to_string(),
                     }
                 )