--- a/src/daft-physical-plan/src/optimization/rules/reorder_partition_keys.rs
+++ b/src/daft-physical-plan/src/optimization/rules/reorder_partition_keys.rs
@@ -156,7 +156,7 @@ impl PhysicalOptimizerRule for ReorderPartitionKeys {
                 PhysicalPlan::InMemoryScan(..) |
                 PhysicalPlan::TabularScan(..) |
                 PhysicalPlan::EmptyScan(..) |
-                PhysicalPlan::PlaceholderScan(..) |
+                PhysicalPlan::PreviousStageScan(..) |
                 PhysicalPlan::Concat(..) |
                 PhysicalPlan::HashJoin(..) |
                 PhysicalPlan::SortMergeJoin(..) |