--- a/src/daft-scheduler/src/scheduler.rs
+++ b/src/daft-scheduler/src/scheduler.rs
@@ -269,6 +269,9 @@ fn physical_plan_to_partition_tasks(
     use daft_dsl::Expr;
     use daft_physical_plan::ops::{CrossJoin, ShuffleExchange, ShuffleExchangeStrategy};
     match physical_plan {
+        PhysicalPlan::PlaceholderScan(..) => {
+            panic!("PlaceholderScan should be optimized away before reaching the scheduler")
+        }
         PhysicalPlan::InMemoryScan(InMemoryScan {
             in_memory_info: InMemoryInfo { cache_key, .. },
             ..