… .github/workflows/python-package.yml
… Cargo.toml
… src/daft-physical-plan/src/ops/mod.rs
… src/daft-physical-plan/src/physical_planner/translate.rs
… src/daft-physical-plan/src/plan.rs
… src/daft-scheduler/src/adaptive.rs
… src/daft-sql/src/lib.rs